Some fixes for 64bit systems
- The usual RemoveItem taking either pointers or int32 and not knowing what to do with 0
- Fix return type for gsloop (int32 instead of long)
- DeviceHandle must be set as a string (-s) instead of a 32bit integer (-d) because it can hold a
  pointer to something in memory (although this is not recommended and deprecated)

+++ src/BGhostview.cpp
@@ -479,7 +479,7 @@ void BGhostview::updateRecentList() {
 	if (!strcmp(filename,(char *) lastOpened->ItemAt(4))) return;
 	int count = m_recent->CountItems();
 	for (int i=0; i<count; i++)
-		delete m_recent->RemoveItem(0l);
+		delete m_recent->RemoveItem((int32)0);
 	void *ofname=lastOpened->LastItem();
 	lastOpened->RemoveItem(ofname);
 	delete ofname;
@@ -836,7 +836,7 @@ void	BGhostview::setup()
 		Lock();
 		int count = marklist->CountItems();
 		for (int i=0; i<count; i++) 
-			delete marklist->RemoveItem(0l);
+			delete marklist->RemoveItem((int32)0);
 		Unlock();
 	};
   if (doc!=NULL) psfree( doc );
